Create vertebra.py
Young's modulus from CT scan is "roughly" working, but needs to be further tested.  Currently it calculates the centroids and maps them back go IJK coordinates and creates a dummy modules based on that calculation.  It appears to work right but needs testing.

This code also creates a grid transform from the displacements but that part needs to be debugged.