Fix grid transform behavior by providing TransformToParent directly
Previously, TransformToParent was computed by inverting
TransformFromParent, leading to numerical instability at transient
singluar areas in the transform.

Now, the transform is directly provided so it can be used without
causing regions to "blow up" at edges.  The transform behavior is
now localized at the domain of the transform.

This change required pre-allocating the vtkGridTransform and the
vtkImageData and setting them into the gridNode before rendering.

We should consider adding these features to slicer.util so that
pre-inverted grid transforms can be added cleanly.